type_inference: move handle_args back to parallel_dispatch to avoid circular imports
