+++ pykokkos/core/type_inference/__init__.py
@@ -1,4 +1,3 @@
 from .args_type_inference import (
-    UpdatedTypes, UpdatedDecorator, HandledArgs,
-    handle_args, get_type_str, get_type_info
+    UpdatedTypes, UpdatedDecorator, get_type_str, get_type_info
 )
